WIP SSO config
